requirements.txt
line-bot-sdk==2.3.0
beautifulsoup4==4.11.1
numpy==1.21.6
requests==2.27.1
flask==2.1.2